add db.Options helper
