+++ db/batch.go
@@ -3,15 +3,11 @@ package db
 import (
 	"context"
 	"database/sql"
-	"fmt"
 	"net/url"
-	"os/user"
 
 	_ "github.com/go-sql-driver/mysql"
-	"github.com/jmoiron/sqlx"
 	_ "github.com/lib/pq"
 	"github.com/pcelvng/task-tools/db/batch"
-	"github.com/pkg/errors"
 )
 
 // BatchLoader implementations should have an initializer that
@@ -53,161 +49,6 @@ type BatchLoader interface {
 	// The order of cols is important and must match the order of row values when
 	// calling AddRow.
 	Commit(ctx context.Context, tableName string, cols ...string) (batch.Stats, error)
-}
-
-// PGx is a convenience initializer to obtain a Postgres sqlx.DB connection
-//
-// Note that this connection will set the default transaction isolation level to
-// 'serializable' to enforce more true atomic batch loading.
-func PGx(user, pass, host, dbName string) (*sqlx.DB, error) {
-	connStr := fmt.Sprintf(
-		"postgres://%s:%s@%s/%s?connect_timeout=5&default_transaction_isolation=serializable&sslmode=disable",
-		user, pass, host, dbName)
-	dbConn, err := sqlx.Connect("postgres", connStr)
-	if err != nil {
-		return nil, err
-	}
-
-	return dbConn, nil
-}
-
-// PGxSSL is a convenience initializer to obtain a Postgres sqlx.DB connection using ssl certs
-//
-// Note that this connection will set the default transaction isolation level to
-// 'serializable' to enforce more true atomic batch loading.
-func PGxSSL(user, pass, host, dbName, sslMode, cert, key, caCert string) (*sqlx.DB, error) {
-	connStr := fmt.Sprintf(
-		"postgres://%s:%s@%s/%s?connect_timeout=5&default_transaction_isolation=serializable&sslmode=%s&sslcert=%s&sslkey=%s&sslrootcert=%s",
-		user, pass, host, dbName, sslMode, cert, key, caCert)
-	dbConn, err := sqlx.Connect("postgres", connStr)
-	if err != nil {
-		return nil, err
-	}
-
-	return dbConn, nil
-}
-
-// PGSSL is a convenience initializer to obtain a Postgres DB connection using ssl certs
-//
-// Note that this connection will set the default transaction isolation level to
-// 'serializable' to enforce more true atomic batch loading.
-func PGSSL(user, pass, host, dbName, sslMode, cert, key, caCert string) (*sql.DB, error) {
-	connStr := fmt.Sprintf(
-		"postgres://%s:%s@%s/%s?connect_timeout=5&default_transaction_isolation=serializable&sslmode=%s&sslcert=%s&sslkey=%s&sslrootcert=%s",
-		user, pass, host, dbName, sslMode, cert, key, caCert)
-	dbConn, err := sql.Open("postgres", connStr)
-	if err != nil {
-		return nil, err
-	}
-
-	// ping
-	if err = dbConn.Ping(); err != nil {
-		return nil, err
-	}
-
-	return dbConn, nil
-}
-
-// MySQL is a convenience initializer to obtain a MySQL DB connection.
-//
-// Note that this connection has an option to set transaction isolation level to
-// 'serializable' to enforce more true atomic batch loading.
-func MySQL(un, pass, host, dbName string) (*sql.DB, error) {
-	connStr := fmt.Sprintf("%s:%s@tcp(%s)/%s?parseTime=true&tx_isolation=serializable", un, pass, host, dbName)
-	dbConn, err := sql.Open("mysql", connStr)
-	if err != nil {
-		return nil, err
-	}
-
-	// ping
-	if err = dbConn.Ping(); err != nil {
-		return nil, err
-	}
-
-	return dbConn, nil
-}
-
-// MySQLTx is a convenience initializer to obtain a MySQL DB connection.
-//
-// Note that this connection will set the default transaction isolation level to
-// 'serializable' to enforce more true atomic batch loading.
-func MySQLTx(un, pass, host, dbName string, serializable bool) (*sql.DB, error) {
-	connStr := fmt.Sprintf("%s:%s@tcp(%s)/%s?parseTime=true", un, pass, host, dbName)
-	if serializable {
-		connStr += "&tx_isolation=serializable"
-	}
-	dbConn, err := sql.Open("mysql", connStr)
-	if err != nil {
-		return nil, err
-	}
-
-	// ping
-	if err = dbConn.Ping(); err != nil {
-		return nil, err
-	}
-
-	return dbConn, nil
-}
-
-// Postgres is a convenience initializer to obtain a Postgres DB connection.
-//
-// Note that this connection will set the default transaction isolation level to
-// 'serializable' to enforce more true atomic batch loading.
-func Postgres(un, pass, host, dbName string) (*sql.DB, error) {
-	if dbName == "" {
-		return nil, errors.New("postgres dbname is required")
-	}
-
-	if un == "" {
-		// postgres user (default is current user)
-		usr, _ := user.Current()
-		un = usr.Username
-	}
-
-	connStr := fmt.Sprintf("postgres://%s:%s@%s/%s?connect_timeout=5&sslmode=disable&default_transaction_isolation=serializable", un, pass, host, dbName)
-	dbConn, err := sql.Open("postgres", connStr)
-	if err != nil {
-		return nil, err
-	}
-
-	// ping
-	if err = dbConn.Ping(); err != nil {
-		return nil, err
-	}
-
-	return dbConn, nil
-}
-
-// PostgresTx is a convenience initializer to obtain a Postgres DB connection
-//
-// Note that this connection has an option to set transaction isolation level to
-// 'serializable' to enforce more true atomic batch loading.
-func PostgresTx(un, pass, host, dbName string, serializable bool) (*sql.DB, error) {
-	if dbName == "" {
-		return nil, errors.New("postgres dbname is required")
-	}
-
-	if un == "" {
-		// postgres user (default is current user)
-		usr, _ := user.Current()
-		un = usr.Username
-	}
-
-	connStr := fmt.Sprintf("postgres://%s:%s@%s/%s?sslmode=disable", un, pass, host, dbName)
-	if serializable {
-		connStr += "&default_transaction_isolation=serializable"
-	}
-	dbConn, err := sql.Open("postgres", connStr)
-	if err != nil {
-		return nil, err
-	}
-
-	// ping
-	if err = dbConn.Ping(); err != nil {
-		return nil, err
-	}
-
-	return dbConn, nil
 }
 
 // NewBatchLoader will create a BatchLoader.
